LayoutTests/platform/chromium-win/fast/css/variables/color-hex-test-expected.txt

CONSOLE MESSAGE: line 27: Uncaught TypeError: Cannot call method 'getVariableValue' of undefined
FAILED
